training pipeline

+++ download_scans.py
@@ -94,7 +94,7 @@ def filter_files(df: pd.DataFrame) -> pd.DataFrame:
     sub_df = df.groupby(["dataset_id", "time"], group_keys=False).apply(
         filter_groups, include_groups=False
     )
-
+    print(sub_df)
     n_valid = sub_df["class"].eq("corrected_pred").sum()
     print(f"Number of valid pairs: {n_valid}")
 
@@ -102,7 +102,10 @@ def filter_files(df: pd.DataFrame) -> pd.DataFrame:
 
 
 @app.command()
-def download_scans(save_dir: Annotated[Path, typer.Argument()]):
+def download_scans(
+    save_dir: Annotated[Path, typer.Argument()],
+    skip_existing: Annotated[bool, typer.Option()],
+):
     omero_client = connect_to_omero()
 
     projects = select_projects_prompt(omero_client)
@@ -122,10 +125,11 @@ def download_scans(save_dir: Annotated[Path, typer.Argument()]):
     for idx, row in tqdm(
         file_subset_df.iterrows(), desc="Downloading Images", total=len(file_subset_df)
     ):
-        image_arr = omero_client.download_image(image_id=row["image_id"])
         file_name = "_".join([row["specimen"], row["time_tag"].lower(), row["class"]])
-
         save_path = os.path.join(save_dir, file_name + ".tiff")
+        if skip_existing and os.path.isfile(save_path):
+            continue
+        image_arr = omero_client.download_image(image_id=row["image_id"])
 
         tiff.imwrite(save_path, image_arr)
